+++ frontend/pages/index.vue
@@ -2,3 +2,7 @@
 const router = useRouter();
 router.push("/bookmarks/");
 </script>
+
+<template>
+  <h1>Welcome!</h1>
+</template>